Перенес запись данных в стейт в main-screen
